refactor: little clean-up
