+++ crates/server/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = "^1.0.65"
 axum = { version = "^0.5.16", features = ["headers"] }
 hyper = { version = "^0.14.20", features = ["full"] }
 serde = { version = "^1.0.145",  features = ["derive"] }
@@ -11,4 +12,5 @@ tokio = { version = "^1.21.2", features = ["full", "tracing"] }
 tracing-subscriber = { version = "^0.3.15", features = ["env-filter"] }
 tower-http = { version = "^0.3.4", features = ["trace"] }
 tower = "^0.4.13"
-tracing = "0.1.36"
+tracing = "^0.1.36"
+futures = "^0.3.24"
