+++ crates/workers/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "^1.0.65"
 tasks = { path = "../tasks" }
 time = "^0.3.9"
 tokio = { version = "^1.21.2", features = ["full", "tracing"] }
